ci: Move srpm/rpm build to packit
Teach make-srpm.sh to support building a directory with the sprm
ingredients (the unpacked directory) to become compatible with the
packit workflow. This mostly obsoletes the need of doing custom COPR
builds, so eventually .copr/ can be simplified. But keep the old
functionality for the time being.

Note that this way of building an srpm in packit is still rather
unusual: the normal mode is to maintain the .spec and all auxiliary
files in the upstream git; that then enables automatic Fedora releases,
and make it easier to keep the spec in sync. But one step after
another..

Configure packit to automatically build srpm and rpms in the usual
temporary COPRs. This makes it much easier to test PRs both by humans
and future integration tests. There are no test plans yet, but already
enable the TF run to at least cover package installation/upgrade.

Run these on Rawhide and the current stable Fedora (38 right now).

This entirely replaces the "build-rpm" workflow, so drop that.